+++ index.js
@@ -3,6 +3,7 @@ module.exports = {
     recommended : require('./configs/recommended'),
   },
   rules : {
-    'align-import' : require('./rules/align-import'),
+    'align-import'     : require('./rules/align-import'),
+    'align-assignment' : require('./rules/align-assignment'),
   },
 };