+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v0.2.0
+
+[compare]()
+
+### 🚀 Features
+
+- Add align-assignment rule
+
+### 🐞 Bug Fixes
+
+- No bug fix, many bugs incoming
+
+### ❤️ Contributors
+
+- Laten Path ([@LatenPath](https://github.com/LatenPath))
+
 ## v0.1.0
 
 [first release](https://github.com/LatenPath/eslint-plugin-perfectionist-align/commit/09deeb2)